Add get_valid_profiles() to Smart Charging Handler.
When retrieving the valid profiles for calculating composite schedules, this function
returns all of the profiles for a given EVSE ID, as well as the station-wide profiles.
It also ensures that the profiles are valid.

The calculations for composite schedule handle the start and end time, so this function
does not account for it at this stage.

Signed-off-by: Christopher Davis <150722105+christopher-davis-afs@users.noreply.github.com>
Signed-off-by: Gianfranco Berardi <54074967+gberardi-pillar@users.noreply.github.com>