smart_charging: Store ChargingLimitSource
When calling `validate_and_add_profile()`, `add_profile()`, and
`insert_or_update_charging_profile()`, we now can be given a source for
the profile. This source is stored in our database and can be retrieved
with the new method, `get_charging_limit_source_for_profile()`.

We use this new method to retrieve the source when retrieving profiles
for K09.

All new functionality is under test.

Co-authored-by: Coury Richards <146002925+couryrr-afs@users.noreply.github.com>
Signed-off-by: Christopher Davis <150722105+christopher-davis-afs@users.noreply.github.com>