smart_charging: Only store and load profiles from db
Remove the map of profiles from the smart charging handler and use the
database for storing, loading, and finding profiles.

We've removed the function for getting all profiles from the
smart charging handler, and in all cases where that was used, including
tests, we retrieve the profiles from the database.

Signed-off-by: Christopher Davis <150722105+christopher-davis-afs@users.noreply.github.com>