smart_charging, database_handler: Use db for K09
Use new functions on the database handler to
implement getting charging profiles by criteria.
This allows us to remove more places where we were
iterating over the entire list of charging
profiles.

This commit includes a fix for a bug where we
were only accepting searches if one of the sources
we were searching for is CSO.

Co-authored-by: Christopher Davis <150722105+christopher-davis-afs@users.noreply.github.com>
Signed-off-by: Coury Richards <146002925+couryrr-afs@users.noreply.github.com>
Signed-off-by: Christopher Davis <150722105+christopher-davis-afs@users.noreply.github.com>