Update ChargePoint tests to use new constructor with dependency injection.
Once all old constructors no longer are used in other areas of the code,
then they can be removed, ensuring the new constructor is used
exclusively.

Also, the tests have been split up to use multiple test fixtures that
cater to the kind of functionality being tested to make it easier to
read and understand for future maintenance.

Signed-off-by: Gianfranco Berardi <54074967+gberardi-pillar@users.noreply.github.com>