smart_charging: Add type for constant limits
When receiving an external limit from another system, we need to know
two things: the limit, and what unit the limit is in. Previously
we only knew the limit (if we're not given a schedule) and made up
a unit.

Now when we receive a constant charging limit we take it as the new type
ConstantChargingLimit, which holds the limit and the unit.

Signed-off-by: Christopher Davis <150722105+christopher-davis-afs@users.noreply.github.com>